Added ReadInt method to Program.Main

+++ AbilityScoreCalculator.cs
@@ -20,7 +20,7 @@ namespace AbilityScoreTester
             double divided = RollResult / DivideBy;
 
             // Add AddAmount to the result of that division
-            int added = AddAmount += divided;
+            int added = AddAmount +=(int)divided;
 
             // If the result is too small, use Minimum
             if (added<Minimum)
